jenkins: more on tracing

+++ src/common/src/get_check_data.c
@@ -190,7 +190,7 @@ double
 pgr_SPI_getFloat8(HeapTuple *tuple, TupleDesc *tupdesc, Column_info_t info) {
     Datum binval;
     bool isnull;
-    float8 value = 0.0;
+    double value = 0.0;
     binval = SPI_getbinval(*tuple, *tupdesc, info.colNumber, &isnull);
     if (isnull)
         elog(ERROR, "Unexpected Null value in column %s", info.name);
